+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update copyright to include 2020
+
 ### Fixed
 * Execute theme generation for theme-aggregator by default, rather than a fall back. This allows applications that pull in old themes and the latest components to correctly contain both theme files.