+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Unreleased
 * Added `TERRA_AGGREGATED_LOCALES` global to the webpack bundle
 
 ### Fixed
-* Re-add log prefix to theme-aggregator
+* Fixed and updated theme-aggregator prefix to [Terra-Toolkit:theme-aggregator] to be consistent with logs from other WDIO services.
 
 5.20.0 - (January 28, 2020)
 ----------